+++ test/index-webgl-dependent-tests.js
@@ -1,7 +1,3 @@
-// Avoid generating a lot of big context divs
-import {setContextDefaults} from '../src/webgl/context';
-setContextDefaults({width: 1, height: 1});
-
 // Imports tests for all modules that depend on webgl
 import './experimental';
 import './deprecated';
